FEAT: Details on pokemons

+++ src/main/java/com/pokebackend/adapter/in/web/PokemonController.java
@@ -1,10 +1,8 @@
 package com.pokebackend.adapter.in.web;
 
 import com.pokebackend.application.service.PokemonService;
+import com.pokebackend.application.service.UserService;
 import com.pokebackend.domain.Pokemon;
-import com.pokebackend.domain.exception.InvalidItemAccessException;
-import com.pokebackend.domain.exception.ItemNotFoundException;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -14,21 +12,22 @@ import java.util.List;
 @RequestMapping("/api/pokemons")
 public class PokemonController {
     private final PokemonService pokemonService;
+    private final UserService userService;
 
-    public PokemonController(PokemonService pokemonService) {
+    public PokemonController(PokemonService pokemonService, UserService userService) {
         this.pokemonService = pokemonService;
+        this.userService = userService;
     }
 
     @PostMapping("/create")
-    public ResponseEntity<Pokemon> createPokemon(@RequestHeader("Authorization") String token,
-                                                 @RequestParam String name,
-                                                 @RequestParam String type,
-                                                 @RequestParam int hp,
-                                                 @RequestParam int attack,
-                                                 @RequestParam int defense,
-                                                 @RequestParam boolean isPublic) {
+    public ResponseEntity<Pokemon> createPokemon(@RequestHeader("Authorization") String token, @RequestParam String name, @RequestParam String type, @RequestParam int hp, @RequestParam int attack, @RequestParam int defense, @RequestParam boolean isPublic) {
+        if (!userService.validateToken(token)) {
+            System.out.println("mal token");
+            return ResponseEntity.status(401).body(null);
+        }
+        System.out.println("bien token");
         Pokemon pokemon = pokemonService.createPokemon(token, name, type, hp, attack, defense, isPublic);
-        return ResponseEntity.status(HttpStatus.CREATED).body(pokemon);
+        return ResponseEntity.ok(pokemon);
     }
 
     @GetMapping("/public")
@@ -39,33 +38,28 @@ public class PokemonController {
 
     @GetMapping("/user")
     public ResponseEntity<List<Pokemon>> getUserPokemons(@RequestHeader("Authorization") String token) {
+        if (!userService.validateToken(token)) {
+            return ResponseEntity.status(401).body(null);
+        }
         List<Pokemon> pokemons = pokemonService.getUserPokemons(token);
         return ResponseEntity.ok(pokemons);
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<Pokemon> updatePokemon(@RequestHeader("Authorization") String token,
-                                                 @PathVariable Long id,
-                                                 @RequestParam String name,
-                                                 @RequestParam String type,
-                                                 @RequestParam int hp,
-                                                 @RequestParam int attack,
-                                                 @RequestParam int defense) {
-        try {
-            Pokemon updatedPokemon = pokemonService.updatePokemon(token, id, name, type, hp, attack, defense);
-            return ResponseEntity.ok(updatedPokemon);
-        } catch (ItemNotFoundException | InvalidItemAccessException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+    public ResponseEntity<Pokemon> updatePokemon(@RequestHeader("Authorization") String token, @PathVariable Long id, @RequestParam String name, @RequestParam String type, @RequestParam int hp, @RequestParam int attack, @RequestParam int defense) {
+        if (!userService.validateToken(token)) {
+            return ResponseEntity.status(401).body(null);
         }
+        Pokemon updatedPokemon = pokemonService.updatePokemon(token, id, name, type, hp, attack, defense);
+        return ResponseEntity.ok(updatedPokemon);
     }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Void> deletePokemon(@RequestHeader("Authorization") String token, @PathVariable Long id) {
-        try {
-            pokemonService.deletePokemon(token, id);
-            return ResponseEntity.noContent().build();
-        } catch (ItemNotFoundException | InvalidItemAccessException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        if (!userService.validateToken(token)) {
+            return ResponseEntity.status(401).build();
         }
+        pokemonService.deletePokemon(token, id);
+        return ResponseEntity.noContent().build();
     }
 }
